Improve dir receive and add an example
Fixes #26

I recently rewrote the zip handling in wormhole-gui for the upcoming 2.2.0 release and I managed to make a few improvements to it at the same time. I figured that I would be a good open source citizen and port it upstream so that everyone can use it.

Most importantly, it makes sure to defer closing of files and does so in a secure way. See https://www.joeshaw.org/dont-defer-close-on-writable-files/. I also lowered a few permissions that were unnecessarily high, an unnecessary type conversion and removed a slow fmt.Sprintf() that could just be a regular string append. I also added the code as an example for the directory send.