Make io.ReadSeeker optional and some code cleanups as well
This commit moves to having SendFile() only take an io.Reader instead of an io.ReadSeeker
and then checks if it conforms to io.ReadSeeker if the WithProgress option is enabled. If
it isn't, it will update the progress in the end, like SendText does. This makes it a lot
more versetile and also allows sending of pretty much any reader. With the benefit of working
a lot better on mobile as well (where your files might not actually be on disk, but instead on iCloud or something).

I also added in a few code cleanups for the send code.